Initial 0.23.x support

+++ src/main/java/org/killbill/billing/plugin/notification/generator/formatters/DefaultInvoiceItemFormatter.java
@@ -201,7 +201,7 @@ public class DefaultInvoiceItemFormatter implements InvoiceItemFormatter {
     }
 
     @Override
-    public Integer getQuantity() {
+    public BigDecimal getQuantity() {
         return item.getQuantity();
     }
